web-analyst-web/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
python-dotenv==1.0.1
openai==1.12.0
faiss-cpu==1.13.2
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.3
readability-lxml==0.8.1
youtube-transcript-api==0.6.2
tiktoken==0.6.0
python-multipart==0.0.6
yt-dlp==2024.3.10
google-api-python-client==2.116.0
google-auth-oauthlib==1.2.0
